backend/requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12

# LangChain ecosystem (1.x stack)
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-google-genai>=4.0.0
langchain-community==0.4.1
langchain-groq>=0.2.0
langchain-chroma>=0.2.0
langgraph>=0.2.0

# Document loaders
pypdf==5.0.1
python-docx==1.1.2
docx2txt==0.8
beautifulsoup4==4.12.3
requests>=2.32.5

# Vector DB
chromadb>=0.6.0

# Utilities
python-dotenv>=1.0.1
tiktoken>=0.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.2.0

# Phase 3 — Action Tools
ddgs>=9.0.0                    # web search, no API key needed (renamed from duckduckgo-search)
tavily-python>=0.3.0           # optional Tavily web search upgrade
